third_party/WebKit/LayoutTests/virtual/mojo-loading/http/tests/inspector/persistence/persistence-tabbed-editor-keeps-selected-tab-expected.txt

Verify that tab keeps selected as the persistence binding comes in.

Opened tabs before persistence binding:
    1: foo.js [selected]
    2: bar.js
    3: foo.js

Opened tabs after persistence binding:
    1: foo.js [selected]
    2: bar.js
